Repurpose is_onscreen() and is_offscreen()
The methods have been renamed to is_contained() and is_outside()
respectively, and will use the passed Surface or Graphic object as
reference instead of supposedly-inherited values for the screen
dimensions.

Part of the reason for this change is that passing the screen
dimensions at runtime would be difficult without creating an
unnecessary instance variable or violating encapsulation principles.
The other reason is that the Rect's x and y-values will be relative
to the destination Surface/Graphic, so it would be impossible to
compare the Graphic's position to the screen anyway.