Graphic: Make position relative to Entity position
Since Graphic is a Component of Entity, its x and y-positions are
simply offsets of the Entity's screen position; it will always be
drawn relative to the Entity's position.

This is an important change, as it allows a single move() function
in Entity (which will be added in a future commit) to 'move' its
Graphic and other positional Components.