Animation: Set _rect.width to the frame width
In most cases, the width of the entire sprite sheet is irrelevant; the
width of one frame is far more important, since it measures what will
actually be shown on-screen.

This also allows the API of Animation to match Graphic's, since most
other methods dealing with width use _rect.width.